+++ webdav-xml/src/properties/displayname.rs
@@ -4,7 +4,7 @@
 
 use bytestring::ByteString;
 
-use crate::{Element, Error, Value, DAV_NAMESPACE, DAV_PREFIX};
+use crate::{Element, ExtractElementError, Value, DAV_NAMESPACE, DAV_PREFIX};
 
 /// The `displayname` property as defined in
 /// [RFC 4918](http://webdav.org/specs/rfc4918.html#PROPERTY_displayname).
@@ -18,7 +18,7 @@ impl Element for DisplayName {
 }
 
 impl TryFrom<&Value> for DisplayName {
-    type Error = Error;
+    type Error = ExtractElementError;
 
     fn try_from(value: &Value) -> Result<Self, Self::Error> {
         Ok(Self(value.to_text()?.clone()))